mydolphin

+++ local-share/icons/mydolphin/mklinks.py
@@ -2,5 +2,6 @@ import os
 for n in os.listdir('/usr/share/icons/breeze'):
     p='/usr/share/icons/breeze/'+n
     if os.path.isdir(p):
-        os.symlink(p,n)
+        if not os.path.exists(n):
+            os.symlink(p,n)
